requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
black==24.8.0
boto3==1.34.158
botocore==1.34.158
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django==5.1
django-extensions==3.2.3
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.7
jmespath==1.0.1
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
